test(#2429): batch 11 — migrate I-group subprocess-core cluster (8 dirs, 273 items) into canonical layers
Full protocol batch: 273 items / 14 glob files, all migrated, zero test
deletions (4 empty __init__.py removed). Layering 251 unit + 22
integration (subprocess/: 559 loopback-WS script renamed off e2e_
prefix + 1395 wrapper cluster; concurrency/: 1005 pause-budget +
1002 heartbeat half).

Hardening: 559 vacuous rejection checks given real outcome assertions
(broad excepts narrowed to WebSocketException/OSError); 1002 dedup
filter swapped to the real production predicate; 1005 wait-routing
hermeticized against build_secure_agent_env (pre-existing main
failure) + assert-message fix; 1461 sys.path leak guarded + parser init
to autouse fixture; 786 pass-throughs strengthened + error assert
pinned; 1395 resolver mocks + user_id=None hermeticity + wrapper
parents[4] anchors; ws_frame monkeypatch.setenv + dead assignment;
markers on all 16 files (+security 1395).

Baselines: issues 2129->1856 / 221->207; default 5960->6233; floor
221->207; inventory 119->111. Local: default 9159+273=9432/653;
issues 2342-273=2069 local (clean 2068/206); unit 7486+251=7737
exact; scanner identical to main (0 new).
