+++ tests/issues/legacy-directories.txt
@@ -63,7 +63,6 @@
 435
 517
 555
-559
 584
 604
 610
@@ -73,7 +72,6 @@
 740
 761
 776
-786
 822
 824
 863
@@ -82,16 +80,12 @@
 913
 993
 1001
-1002
-1005
 1006
 1138
 1140
 1306
 1329
-1395
 1425
-1461
 1469
 1647
 1746
@@ -114,6 +108,4 @@
 2183
 2189
 2280
-2339
-2340
 2376
